Set up a right/wrong button page
This is to be used while playing a game where the random Chinese
functionality gives a word, the student makes a sentence with the
word, and the teacher can indicate whether the student's sentence
was right or wrong. At some point the random page and the buttons
page would be connected (the "right" button advances to the next
random choice).